+++ Task1_Part2/Task1_Part2/Matrixs.cpp
@@ -11,7 +11,7 @@ SMatrixs::SMatrixs()
 		basicMatrix[i].resize(size);
 		for (size_t j = 0; j != size; ++j)
 		{
-			basicMatrix[i][j] = rand() % 9 + 1;
+			basicMatrix[i][j] = static_cast<float>(rand() % 9 + 1);
 		}
 	}
 	minorsMatrix.resize(size);
